requirements.txt
alembic
flask
Flask-Migrate==1.3.0
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
Werkzeug
Jinja2
itsdangerous
markupsafe
requests
gunicorn
psycopg2==2.5.4
